refactor(opencode): isolate compaction routing
