+++ server-agents/opencode/src/agents/opencode/compaction-routing.ts
@@ -1,0 +1,87 @@
+import type { AgentLogger } from '@garcon/server-agent-interface';
+import type {
+  OpenCodeOperationRoute,
+  OpenCodeOperationRoutes,
+} from './operation-routes.js';
+import type { SSEEvent } from './sse-events.js';
+import type { OpenCodeSession } from './turn-events.js';
+
+type OpenCodeCompactionPartDropCode =
+  | 'COMPACTION_PART_NO_SESSION'
+  | 'COMPACTION_PART_SESSION_NOT_RUNNING'
+  | 'COMPACTION_PART_BEFORE_PROMPT'
+  | 'COMPACTION_PART_ROUTE_RETIRED'
+  | 'COMPACTION_PART_INVALID_IDENTIFIERS'
+  | 'COMPACTION_PART_IDENTITY_COLLISION';
+
+interface OpenCodeCompactionPartRouteInput {
+  readonly event: SSEEvent;
+  readonly logger: AgentLogger;
+  readonly operationRoutes: OpenCodeOperationRoutes;
+  readonly session: OpenCodeSession | undefined;
+  readonly sessionId: string;
+}
+
+export function adoptOpenCodeCompactionPartRoute(
+  input: OpenCodeCompactionPartRouteInput,
+): OpenCodeOperationRoute | null {
+  const { event, logger, operationRoutes, session, sessionId } = input;
+  if (!session) return dropCompactionPart(logger, 'COMPACTION_PART_NO_SESSION', sessionId, event);
+  if (session.status !== 'running') {
+    return dropCompactionPart(
+      logger,
+      'COMPACTION_PART_SESSION_NOT_RUNNING',
+      sessionId,
+      event,
+    );
+  }
+  if (session.turn.providerMessageId === null) {
+    return dropCompactionPart(logger, 'COMPACTION_PART_BEFORE_PROMPT', sessionId, event);
+  }
+
+  const adoption = operationRoutes.adoptCompactionPart(session.turn, event);
+  switch (adoption.kind) {
+    case 'adopted': {
+      const part = event.properties?.part;
+      logger.debug('Adopted an OpenCode compaction part', {
+        agentSessionId: sessionId,
+        partId: typeof part?.id === 'string' ? part.id : null,
+        messageId: typeof part?.messageID === 'string' ? part.messageID : null,
+      });
+      return adoption.route;
+    }
+    case 'route-retired':
+      return dropCompactionPart(logger, 'COMPACTION_PART_ROUTE_RETIRED', sessionId, event);
+    case 'invalid-identifiers':
+      return dropCompactionPart(
+        logger,
+        'COMPACTION_PART_INVALID_IDENTIFIERS',
+        sessionId,
+        event,
+      );
+    case 'identity-collision':
+      return dropCompactionPart(
+        logger,
+        'COMPACTION_PART_IDENTITY_COLLISION',
+        sessionId,
+        event,
+      );
+  }
+}
+
+function dropCompactionPart(
+  logger: AgentLogger,
+  code: OpenCodeCompactionPartDropCode,
+  sessionId: string,
+  event: SSEEvent,
+): null {
+  const part = event.properties?.part;
+  logger.warn('Dropping an OpenCode compaction part', {
+    code,
+    agentSessionId: sessionId,
+    eventId: event.id ?? null,
+    partId: typeof part?.id === 'string' ? part.id : null,
+    messageId: typeof part?.messageID === 'string' ? part.messageID : null,
+  });
+  return null;
+}
